Re-organizing some files on disk to make it more obvious when I'm sharing between demos, switched the demos to depend on JLFGameplayKit libraries, instead of directly on source.

+++ Common/BouncyComponent.h
@@ -8,13 +8,13 @@
 
 #import "JLFGKComponent.h"
 
-extern const CGFloat kDefaultBounceSpeed;
-extern const CGFloat kDefaultBaseAnchorHeight;
+extern const float kDefaultBounceSpeed;
+extern const float kDefaultBaseAnchorHeight;
 
 @interface BouncyComponent : JLFGKComponent
 
-@property (assign, nonatomic) CGFloat bounceSpeed;
-@property (assign, nonatomic) CGFloat baseAnchorHeight;
+@property (assign, nonatomic) float bounceSpeed;
+@property (assign, nonatomic) float baseAnchorHeight;
 
 - (void)startBouncing;
 - (void)stopBouncing;
